fix: wrong API path, fix wrong format input
